feat: adiciona api que lista professores, salas e palavras

+++ src/main/java/com/tupinet/games/repository/PontuacaoRepository.java
@@ -16,5 +16,5 @@ public interface PontuacaoRepository extends JpaRepository<Pontuacao, Integer> {
             "JOIN sj.jogo j " +
             "JOIN sj.sala s " +
             "WHERE s.id = :salaId")
-    List<RankingDTO> findPontuacaoBySalaId(@Param("salaId") Long salaId);
+    List<RankingDTO> findPontuacaoBySalaId(@Param("salaId") Integer salaId);
 }
